# Optimized the collapsing of PointSources