--- a/openquake/hazardlib/source/point.py
+++ b/openquake/hazardlib/source/point.py
@@ -214,12 +214,13 @@ class PointSource(ParametricSeismicSource):
             arr['dims'] = get_rupdims(areas, np.dip, width, rar)
         return planin
 
-    def max_radius(self):
+    def max_radius(self, maxdist):
         """
         :returns: max radius + ps_grid_spacing * sqrt(2)/2
         """
-        max_rp_radius = self._get_max_rupture_projection_radius()
-        return self.ps_grid_spacing * .707 + max_rp_radius
+        self._get_max_rupture_projection_radius()
+        eff_radius = min(self.radius[-1], maxdist / 2)
+        return eff_radius + self.ps_grid_spacing * .707
 
     def get_psdist(self, m, mag, psdist, magdist):
         """
@@ -354,8 +355,8 @@ class PointSource(ParametricSeismicSource):
         """
         Bounding box of the point, enlarged by the maximum distance
         """
-        delta = min(self.max_radius(), maxdist)
-        return get_bounding_box([self.location], maxdist + delta)
+        radius = self.max_radius(maxdist)
+        return get_bounding_box([self.location], maxdist + radius)
 
     def wkt(self):
         """