--- a/openquake/qa_tests_data/classical/case_43/expected/hazard_map-mean-PGA.csv
+++ b/openquake/qa_tests_data/classical/case_43/expected/hazard_map-mean-PGA.csv
@@ -1,3 +1,3 @@
-#,,"generated_by='OpenQuake engine 3.15.0-git4f40f27632', start_date='2022-06-09T08:24:24', checksum=4118480633, kind='mean', investigation_time=50.0"
+#,,"generated_by='OpenQuake engine 3.16.0-git4c75d2fe4c', start_date='2022-11-08T03:23:56', checksum=4118480633, kind='mean', investigation_time=50.0"
 lon,lat,PGA-0.002105
-112.10000,22.30000,3.189432E-02
+112.10000,22.30000,3.188878E-02