… openquake/qa_tests_data/disagg/case_4/expected_output/Mag_Dist-0.csv
… openquake/qa_tests_data/disagg/case_4/expected_output/Mag_Dist_TRT-0.csv