With a very minor change in the numbers we can get a decent speedup. Here are the figures for the new AUS model (reduced):
```
# before
| calc_48862, maxmem=98.8 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 44_233   | 318.9     | 253       |
| get_poes                   | 12_914   | 0.0       | 6_320_704 |
| computing mean_std         | 12_366   | 0.0       | 245_756   |
| composing pnes             | 9_148    | 0.0       | 6_320_704 |
| planar contexts            | 4_505    | 0.0       | 258_630   |
| collapsing contexts        | 3_387    | 356.4     | 4_834     |
| ClassicalCalculator.run    | 512.8    | 595.5     | 1         |
# after
| calc_48863, maxmem=96.6 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 28_413   | 309.0     | 253       |
| get_poes                   | 8_262    | 0.0       | 4_096_032 |
| computing mean_std         | 7_892    | 0.0       | 159_373   |
| composing pnes             | 5_939    | 0.0       | 4_096_032 |
| planar contexts            | 2_934    | 0.0       | 262_008   |
| collapsing contexts        | 2_116    | 236.1     | 3_198     |
| ClassicalCalculator.run    | 336.3    | 568.0     | 1         |
```
Hopefully, by choosing carefully the parameters, the new AUS model can now be done in 2-3 days instead of 4 days.